+++ packages/webkit/src/components/avatar/avatar.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
   import { computed, useAttrs } from 'vue'
 
-  import { cn } from '../../../utils/cn'
+  import { cn } from '../../utils/cn'
 
   export type AvatarKind = 'circle' | 'square'
   export type AvatarSize = 'small' | 'medium' | 'large'